Refactor CTResponse
This patch refactors CTResponse class
